+++ packages/react-i18n/src/test/utilities.test.tsx
@@ -108,6 +108,16 @@ describe('getTranslationTree()', () => {
     ).toBe('other');
   });
 
+  it('returns the translation keys even if they are hidden in a second dictionary', () => {
+    expect(
+      getTranslationTree(
+        'foobar.barfoo',
+        [{foo: {bar: 'one'}}, {foobar: {barfoo: 'two'}}],
+        locale,
+      ),
+    ).toBe('two');
+  });
+
   it('throws a MissingTranslationError when no translation is found', () => {
     expect(() =>
       getTranslationTree('foo.bar.baz', {foo: {bar: 'one'}}, locale),